Added help command
A very basic help command to print out help messages for subcommands has been added. The command can be registered at any point and takes a pointer to a vector of pointers to all other commands.

An update to the global help has also been made to include a note directing the user to use the help command for command specific help.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>