Initial implementation of extractor
This is the initial implementation of the extractor. We probably don't need the progress stream buf as we can use the getBytesRead instead. Profiling shows that the majority of the time is spent on the bzip decompression. It might therefore be worth testing with a pre-decompressed file or with the 7z files instead.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>