Moved EnsureArgument to base command. Added read message to base command
Moved EnsureArgument to the base command because it is useful across commands, not just in extract.

ReadMessage is a useful one across all commands because chances are we have to read a message pretty much everywhere. It is probably worth moving write message as well but we can do that when it comes to it.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>